Combined array field defs.
